Move db setup functions to a separate module
